Refactor review to folder, add pagination
