+++ core/static/core/scss/components/pagination.scss
@@ -20,3 +20,8 @@
     margin-right: 0.5em;
     cursor: pointer;
 }
+
+.page-item {
+    display: inline-block;
+
+}
