+++ core/templatetags/cl_filters.py
@@ -1,4 +1,5 @@
 from django import template
+from django.core.paginator import Page
 
 register = template.Library()
 
@@ -11,3 +12,8 @@ def reverse_list(values):
 @register.filter
 def is_general_true(value):
     return value in (1, '1', True, 'Y', 'y',)
+
+
+@register.filter
+def get_elided_page_range(page: Page, on_each_side=3, on_ends=2):
+    return page.paginator.get_elided_page_range(number=page.number, on_each_side=on_each_side, on_ends=on_ends)
