+++ uploader/review/review.py
@@ -1,0 +1,189 @@
+import logging
+from datetime import datetime
+from typing import List, Type
+
+from django.core.exceptions import FieldDoesNotExist
+from django.db import IntegrityError, models
+from django.db.models import QuerySet
+
+from core.constant import REL_TYPE_STORED_IN, REL_TYPE_CREATED, REL_TYPE_WAS_ADDRESSED_TO, \
+    REL_TYPE_PEOPLE_MENTIONED_IN_WORK, REL_TYPE_WAS_SENT_TO, REL_TYPE_WAS_SENT_FROM, REL_TYPE_IS_RELATED_TO
+from core.models import CofkUnionResource
+from institution.models import CofkUnionInstitution
+from manifestation.models import CofkUnionManifestation, CofkManifInstMap
+from uploader.models import CofkCollectUpload
+from work.models import CofkCollectWork, CofkUnionWork, CofkWorkLocationMap, CofkWorkPersonMap, CofkWorkResourceMap, \
+    CofkUnionLanguageOfWork
+
+
+log = logging.getLogger(__name__)
+
+
+def create_union_work(collect_work: CofkCollectWork):
+    union_dict = {
+        # work_id is primary key in CofkUnionWork
+        'work_id': f'work_{datetime.now().strftime("%Y%m%d%H%M%S%f")}_{collect_work.iwork_id}',
+    }
+
+    for field in [f for f in collect_work._meta.get_fields() if f.name != 'iwork_id']:
+        try:
+            CofkUnionWork._meta.get_field(field.name)
+
+            if value := getattr(collect_work, field.name):
+                union_dict[field.name] = value
+
+        except FieldDoesNotExist:
+            # log.warning(f'Field {field} does not exist')
+            pass
+
+    return CofkUnionWork(**union_dict)
+
+
+def link_person_to_work(entities: QuerySet, relationship_type: str, union_work: CofkUnionWork, work_id, request) \
+        -> List[CofkWorkPersonMap]:
+    person_maps = []
+    for person in entities.filter(iwork_id=work_id).all():
+        cwpm = CofkWorkPersonMap(relationship_type=relationship_type,
+                                 work=union_work, person=person.iperson.union_iperson,
+                                 person_id=person.iperson.union_iperson.person_id)
+        cwpm.update_current_user_timestamp(request.user.username)
+        # cwpm.save()
+        person_maps.append(cwpm)
+
+    return person_maps
+
+
+def link_location_to_work(entities: QuerySet, relationship_type: str, union_work: CofkUnionWork, work_id, request) \
+        -> List[CofkWorkLocationMap]:
+    location_maps = []
+    for origin_or_dest in entities.filter(iwork_id=work_id).all():
+        cwlm = CofkWorkLocationMap(relationship_type=relationship_type,
+                                   work=union_work, location=origin_or_dest.location.union_location,
+                                   location_id=origin_or_dest.location.union_location.location_id)
+        cwlm.update_current_user_timestamp(request.user.username)
+        # cwlm.save()
+        location_maps.append(cwlm)
+
+    return location_maps
+
+
+def create_union_manifestations(work_id: int, union_work: CofkUnionWork, request, context) -> List[CofkManifInstMap]:
+    union_maps = []
+    for manif in context['manifestations'].filter(iwork_id=work_id).all():
+        union_dict = {'manifestation_creation_date_is_range': 0}
+        for field in [f for f in manif._meta.get_fields() if f.name != 'iwork_id']:
+            try:
+                CofkUnionManifestation._meta.get_field(field.name)
+                union_dict[field.name] = getattr(manif, field.name)
+
+            except FieldDoesNotExist:
+                # log.warning(f'Field {field} does not exist')
+                pass
+
+        union_manif = CofkUnionManifestation(**union_dict)
+        union_manif.work = union_work
+        union_manif.save()
+
+        if manif.repository_id is not None:
+            inst = context['institutions'].filter(id=manif.repository_id).first()
+            union_inst = CofkUnionInstitution.objects.filter(pk=inst.institution_id).first()
+
+            cmim = CofkManifInstMap(relationship_type=REL_TYPE_STORED_IN,
+                                    manif=union_manif, inst=union_inst, inst_id=union_inst.institution_id)
+            cmim.update_current_user_timestamp(request.user.username)
+            union_maps.append(cmim)
+            # cmim.save()
+
+    return union_maps
+
+
+def bulk_create(objects: List[Type[models.Model]]):
+    if objects:
+        try:
+            type(objects[0]).objects.bulk_create(objects, batch_size=500)
+        except IntegrityError as ie:
+            log.error(ie)
+            log.exception(ie)
+            # self.add_error(f'Could not create {type(objects[0])} objects in database.')
+
+
+def accept_work(request, context: dict, upload: CofkCollectUpload):
+    try:
+        work_id = int(request.GET['work_id'])
+    except ValueError:
+        return
+    collect_work = [w for w in context['works_page'].object_list if w.id == work_id][0]
+
+    if collect_work.upload_status_id != 1:
+        return
+
+    # Create work
+    union_work = create_union_work(collect_work)
+    union_work.save()
+
+    # Link people
+    people_maps = link_person_to_work(entities=context['authors'], relationship_type=REL_TYPE_CREATED,
+                                      union_work=union_work, work_id=work_id, request=request)
+    people_maps += link_person_to_work(entities=context['addressees'], relationship_type=REL_TYPE_WAS_ADDRESSED_TO,
+                                       union_work=union_work, work_id=work_id, request=request)
+    people_maps += link_person_to_work(entities=context['mentioned'],
+                                       relationship_type=REL_TYPE_PEOPLE_MENTIONED_IN_WORK,
+                                       union_work=union_work, work_id=work_id, request=request)
+    bulk_create(people_maps)
+
+    lang_maps = [CofkUnionLanguageOfWork(work=union_work, language_code=lang.language_code) for
+                 lang in context['languages'].filter(iwork_id=work_id).all()]
+    # Link languages
+    bulk_create(lang_maps)
+
+    # Link locations
+    loc_maps = link_location_to_work(entities=context['destinations'], relationship_type=REL_TYPE_WAS_SENT_TO,
+                          union_work=union_work, work_id=work_id, request=request)
+    loc_maps += link_location_to_work(entities=context['origins'], relationship_type=REL_TYPE_WAS_SENT_FROM,
+                          union_work=union_work, work_id=work_id, request=request)
+    bulk_create(loc_maps)
+
+    # Create manifestations
+    union_maps = create_union_manifestations(work_id=work_id, union_work=union_work, request=request, context=context)
+    bulk_create(union_maps)
+
+    # Link resources
+    for resource in context['resources'].filter(iwork_id=work_id).all():
+        union_resource = CofkUnionResource()
+        union_resource.resource_url = resource.resource_url
+        union_resource.resource_name = resource.resource_name
+        union_resource.resource_details = resource.resource_details
+        union_resource.resource_id = resource.resource_id
+        union_resource.save()
+
+        cwrm = CofkWorkResourceMap(relationship_type=REL_TYPE_IS_RELATED_TO, work=union_work,
+                                   resource=union_resource, resource_id=union_resource.resource_id)
+        cwrm.update_current_user_timestamp(request.user.username)
+        cwrm.save()
+
+    # Change state of upload and work
+    upload.upload_status_id = 2  # Partly reviewed
+    upload.works_accepted += 1
+    upload.save()
+
+    collect_work.upload_status_id = 4  # Accepted and saved into main database
+    collect_work.save()
+
+
+def reject_work(request, context: dict, upload: CofkCollectUpload):
+    work_id = request.GET['work_id']
+    collect_work = context['works'].filter(pk=work_id).first()
+
+    if collect_work.upload_status_id != 1:
+        return
+
+    upload.upload_status_id = 2  # Partly reviewed
+    upload.works_rejected += 1
+    # upload.save()
+
+    collect_work.upload_status_id = 5  # Rejected
+    # collect_work.save()
+
+
+def accept_works(request, context: dict, upload: CofkCollectUpload):
+    pass
