+++ uploader/templates/uploader/review.html
@@ -8,20 +8,23 @@
     <span id="upload-summary">Status: {{ upload.upload_status }} | Number of works uploaded: {{ upload.total_works }} | Accepted: {{ upload.works_accepted}} | Rejected: {{ upload.works_rejected}}</span>
 
     <div id="upload-options">
+    <form>
     Back to list of contributions <button type="button" class="btn inline_btn" onclick="location.href='/upload'">Back</button>
 
-    Accept entire contribution <button class="btn inline_btn" disabled>Accept all</button>
 
-    Reject entire contribution <button class="btn inline_btn" disabled>Reject all</button>
+    Accept entire contribution <button class="btn inline_btn" name="accept_all">Accept all</button>
+
+    Reject entire contribution <button class="btn inline_btn" name="reject_all">Reject all</button>
+    </form>
     </div>
     <em>Note: confirmation will be required before Accept/Reject of entire contribution.</em>
 </div>
 
 {% include "uploader/component/works.html" %}
-
+<!--
 {% include "uploader/component/people.html" %}
 
-<!--
+
 
 {% for location in places %}
