+++ uploader/views.py
@@ -1,37 +1,27 @@
 import logging
 import os
 import time
-from datetime import datetime
 from zipfile import BadZipFile
 
 from django.contrib.auth.decorators import login_required
-from django.core.exceptions import FieldDoesNotExist
 from django.core.files.storage import default_storage
 from django.core.paginator import Paginator
-from django.db.models import QuerySet
 from django.shortcuts import render, redirect
 from django.utils import timezone
-from pandas._config.config import OptionError
 
-from core.constant import REL_TYPE_CREATED, REL_TYPE_WAS_ADDRESSED_TO, REL_TYPE_PEOPLE_MENTIONED_IN_WORK, \
-    REL_TYPE_WAS_SENT_TO, REL_TYPE_WAS_SENT_FROM, REL_TYPE_STORED_IN, REL_TYPE_IS_RELATED_TO
-from core.models import CofkUnionResource
-from institution.models import CofkCollectInstitution, CofkUnionInstitution
+from institution.models import CofkCollectInstitution
 from location.models import CofkCollectLocation
-from manifestation.models import CofkCollectManifestation, CofkUnionManifestation, CofkManifInstMap
+from manifestation.models import CofkCollectManifestation
 from person.models import CofkCollectPerson
 from uploader.forms import CofkCollectUploadForm
 from django.conf import settings
 
-from uploader.models import CofkCollectStatus, CofkCollectUpload
+from uploader.models import CofkCollectUpload
+from uploader.review.review import accept_work, reject_work, accept_works
 from uploader.spreadsheet import CofkUploadExcelFile
 from uploader.validation import CofkMissingColumnError, CofkMissingSheetError, CofkNoDataError
 
-from work.models import CofkCollectWork, CofkCollectAuthorOfWork, CofkCollectAddresseeOfWork, CofkCollectLanguageOfWork, \
-    CofkCollectPersonMentionedInWork, CofkCollectWorkResource, CofkUnionWork, CofkWorkPersonMap, \
-    CofkCollectDestinationOfWork, CofkCollectOriginOfWork, CofkWorkLocationMap, CofkUnionLanguageOfWork, \
-    CofkWorkResourceMap
-
+from work.models import CofkCollectWork, CofkCollectAuthorOfWork, CofkCollectAddresseeOfWork, CofkCollectLanguageOfWork, CofkCollectPersonMentionedInWork, CofkCollectWorkResource, CofkCollectDestinationOfWork, CofkCollectOriginOfWork
 log = logging.getLogger(__name__)
 
 
@@ -42,7 +32,8 @@ def handle_upload(request, context):
 
         start = time.time()
         new_upload = form.save(commit=False)
-        new_upload.upload_status = CofkCollectStatus.objects.filter(status_id=1).first()
+        new_upload.upload_status_id = 1
+        new_upload.uploader_email = request.user.email
         new_upload.upload_timestamp = timezone.now()
         new_upload.save()
 
@@ -88,7 +79,7 @@ def handle_upload(request, context):
             context['report']['errors'] = {'file': {'total': len(errors), 'error': errors}}
             # log.error(ve.args[0])
             log.error([err for i, err in enumerate(cmce.args[0]) if i % 2 != 0])
-        except (FileNotFoundError, BadZipFile, OptionError, OSError) as e:
+        except (FileNotFoundError, BadZipFile, OSError) as e:
             context['report']['total_errors'] = 1
             context['report']['errors'] = {'file': {'total': 1, 'error': ['Could not read the Excel file.']}}
             log.error(e)
@@ -139,148 +130,6 @@ def upload_view(request, **kwargs):
     return render(request, template_url, context)
 
 
-def create_union_work(collect_work: CofkCollectWork):
-    union_dict = {
-        # work_id is primary key in CofkUnionWork
-        'work_id': f'work_{datetime.now().strftime("%Y%m%d%H%M%S%f")}_{collect_work.iwork_id}',
-    }
-
-    for field in [f for f in collect_work._meta.get_fields() if f.name != 'iwork_id']:
-        try:
-            CofkUnionWork._meta.get_field(field.name)
-
-            if value := getattr(collect_work, field.name):
-                union_dict[field.name] = value
-
-        except FieldDoesNotExist:
-            # log.warning(f'Field {field} does not exist')
-            pass
-
-    union_work = CofkUnionWork(**union_dict)
-    union_work.save()
-
-    return union_work
-
-
-def link_person_to_work(entities: QuerySet, relationship_type: str,
-                        union_work: CofkUnionWork, work_id, request):
-    for person in entities.filter(iwork_id=work_id).all():
-        cwpm = CofkWorkPersonMap(relationship_type=relationship_type,
-                                 work=union_work, person=person.iperson.union_iperson,
-                                 person_id=person.iperson.union_iperson.person_id)
-        cwpm.update_current_user_timestamp(request.user.username)
-        cwpm.save()
-
-
-def link_location_to_work(entities: QuerySet, relationship_type: str,
-                          union_work: CofkUnionWork, work_id, request):
-    for origin_or_dest in entities.filter(iwork_id=work_id).all():
-        cwlm = CofkWorkLocationMap(relationship_type=relationship_type,
-                                   work=union_work, location=origin_or_dest.location.union_location,
-                                   location_id=origin_or_dest.location.union_location.location_id)
-        cwlm.update_current_user_timestamp(request.user.username)
-        cwlm.save()
-
-
-def create_union_manifestations(work_id: str, union_work: CofkUnionWork, request, context):
-    for manif in context['manifestations'].filter(iwork_id=work_id).all():
-        union_dict = {'manifestation_creation_date_is_range': 0}
-        for field in [f for f in manif._meta.get_fields() if f.name != 'iwork_id']:
-            try:
-                CofkUnionManifestation._meta.get_field(field.name)
-                union_dict[field.name] = getattr(manif, field.name)
-
-            except FieldDoesNotExist:
-                # log.warning(f'Field {field} does not exist')
-                pass
-
-        union_manif = CofkUnionManifestation(**union_dict)
-        union_manif.work = union_work
-        union_manif.save()
-
-        if manif.repository_id is not None:
-            inst = context['institutions'].filter(id=manif.repository_id).first()
-            union_inst = CofkUnionInstitution.objects.filter(pk=inst.institution_id).first()
-
-            cmim = CofkManifInstMap(relationship_type=REL_TYPE_STORED_IN,
-                                    manif=union_manif, inst=union_inst, inst_id=union_inst.institution_id)
-            cmim.update_current_user_timestamp(request.user.username)
-            cmim.save()
-
-
-def accept_work(request, context: dict, upload: CofkCollectUpload):
-    try:
-        work_id = int(request.GET['work_id'])
-    except ValueError:
-        return
-    collect_work = [w for w in context['works_page'].object_list if w.id == work_id][0]
-
-    if collect_work.upload_status_id != 1:
-        return
-
-    # Create work
-    union_work = create_union_work(collect_work)
-
-    # Link people
-    link_person_to_work(entities=context['authors'], relationship_type=REL_TYPE_CREATED,
-                        union_work=union_work, work_id=work_id, request=request)
-    link_person_to_work(entities=context['addressees'], relationship_type=REL_TYPE_WAS_ADDRESSED_TO,
-                        union_work=union_work, work_id=work_id, request=request)
-    link_person_to_work(entities=context['mentioned'], relationship_type=REL_TYPE_PEOPLE_MENTIONED_IN_WORK,
-                        union_work=union_work, work_id=work_id, request=request)
-
-    # Link languages
-    for lang in context['languages'].filter(iwork_id=work_id).all():
-        CofkUnionLanguageOfWork(work=union_work, language_code=lang.language_code).save()
-
-    # Link locations
-    link_location_to_work(entities=context['destinations'], relationship_type=REL_TYPE_WAS_SENT_TO,
-                          union_work=union_work, work_id=work_id, request=request)
-    link_location_to_work(entities=context['origins'], relationship_type=REL_TYPE_WAS_SENT_FROM,
-                          union_work=union_work, work_id=work_id, request=request)
-
-    # Create manifestations
-    create_union_manifestations(work_id=work_id, union_work=union_work,
-                                request=request, context=context)
-
-    # Link resources
-    for resource in context['resources'].filter(iwork_id=work_id).all():
-        union_resource = CofkUnionResource()
-        union_resource.resource_url = resource.resource_url
-        union_resource.resource_name = resource.resource_name
-        union_resource.resource_details = resource.resource_details
-        union_resource.resource_id = resource.resource_id
-        union_resource.save()
-
-        cwrm = CofkWorkResourceMap(relationship_type=REL_TYPE_IS_RELATED_TO,
-                                   work=union_work, resource=union_resource, resource_id=union_resource.resource_id)
-        cwrm.update_current_user_timestamp(request.user.username)
-        cwrm.save()
-
-    # Change state of upload and work
-    upload.upload_status_id = 2  # Partly reviewed
-    upload.works_accepted += 1
-    upload.save()
-
-    collect_work.upload_status_id = 4  # Accepted and saved into main database
-    collect_work.save()
-
-
-def reject_work(request, context: dict, upload: CofkCollectUpload):
-    work_id = request.GET['work_id']
-    collect_work = context['works'].filter(pk=work_id).first()
-
-    if collect_work.upload_status_id != 1:
-        return
-
-    upload.upload_status_id = 2  # Partly reviewed
-    upload.works_rejected += 1
-    # upload.save()
-
-    collect_work.upload_status_id = 5  # Rejected
-    # collect_work.save()
-
-
 @login_required
 def upload_review(request, upload_id, **kwargs):
     template_url = 'uploader/review.html'
@@ -311,5 +160,7 @@ def upload_review(request, upload_id, **kwargs):
             accept_work(request, context, upload)
         elif 'reject_work' in request.GET:
             reject_work(request, context, upload)
+        elif 'accept_works' in request.GET:
+            accept_works(request, context, upload)
 
     return render(request, template_url, context)
